refactor: move privacy policy fetch service to namespace Telemetry.PrivacyPolicy
